Rename externals package to utils
